Crate nautilus_common

PoseiTrader is an open-source, high-performance, production-grade algorithmic trading platform, providing quantitative traders with the ability to backtest portfolios of automated trading strategies on historical data with an event-driven engine, and also deploy those same strategies live, with no code changes.

Feature flags

This crate provides feature flags to control source code inclusion during compilation, depending on the intended use case, i.e. whether to provide Python bindings for the posei_trader Python package, or as part of a Rust only build.

ffi: Enables the C foreign function interface (FFI) from cbindgen.

python: Enables Python bindings from PyO3.

stubs: Enables type stubs for use in testing scenarios.

Modules

actor

cache

In-memory cache for market and execution data, with optional persistent backing.

clock

Real-time and static test Clock implementations.

component

custom

A user custom data type.

enums

Enumerations for common components.

factories

Factories for constructing domain objects such as orders.

ffi

C foreign function interface (FFI) from cbindgen.

generators

Provides generation of identifiers such as ClientOrderId and PositionId.

greeks

Greeks calculator for options and futures.

logging

The logging framework for Posei systems.

messages

msgbus

A common in-memory MessageBus supporting multiple messaging patterns:

python

Python bindings from PyO3.

runner

runtime

The centralized Tokio runtime for a running Posei system.

signal

A user signal type.

testing

Common test related helper functions.

throttler

timer

Real-time and test timers for use with Clock implementations.

xrate

Exchange rate calculations between currencies.