nautilus_common::actor::data_actor

Trait DataActor

pub trait DataActor:
    Component
    + Deref<Target = DataActorCore>
    + DerefMut<Target = DataActorCore> {

    // Provided methods
    fn on_save(&self) -> Result<IndexMap<String, Vec<u8>>> { ... }
    fn on_load(&mut self, state: IndexMap<String, Vec<u8>>) -> Result<()> { ... }
    fn on_start(&mut self) -> Result<()> { ... }
    fn on_stop(&mut self) -> Result<()> { ... }
    fn on_resume(&mut self) -> Result<()> { ... }
    fn on_reset(&mut self) -> Result<()> { ... }
    fn on_dispose(&mut self) -> Result<()> { ... }
    fn on_degrade(&mut self) -> Result<()> { ... }
    fn on_fault(&mut self) -> Result<()> { ... }
    fn on_time_event(&mut self, event: &TimeEvent) -> Result<()> { ... }
    fn on_data(&mut self, data: &dyn Any) -> Result<()> { ... }
    fn on_signal(&mut self, signal: &Signal) -> Result<()> { ... }
    fn on_instrument(&mut self, instrument: &InstrumentAny) -> Result<()> { ... }
    fn on_book_deltas(&mut self, deltas: &OrderBookDeltas) -> Result<()> { ... }
    fn on_book(&mut self, order_book: &OrderBook) -> Result<()> { ... }
    fn on_quote(&mut self, quote: &QuoteTick) -> Result<()> { ... }
    fn on_trade(&mut self, tick: &TradeTick) -> Result<()> { ... }
    fn on_bar(&mut self, bar: &Bar) -> Result<()> { ... }
    fn on_mark_price(&mut self, mark_price: &MarkPriceUpdate) -> Result<()> { ... }
    fn on_index_price(&mut self, index_price: &IndexPriceUpdate) -> Result<()> { ... }
    fn on_instrument_status(&mut self, data: &InstrumentStatus) -> Result<()> { ... }
    fn on_instrument_close(&mut self, update: &InstrumentClose) -> Result<()> { ... }
    fn on_block(&mut self, block: &Block) -> Result<()> { ... }
    fn on_pool(&mut self, pool: &Pool) -> Result<()> { ... }
    fn on_pool_swap(&mut self, swap: &PoolSwap) -> Result<()> { ... }
    fn on_pool_liquidity_update(
        &mut self,
        update: &PoolLiquidityUpdate,
    ) -> Result<()> { ... }
    fn on_historical_data(&mut self, data: &dyn Any) -> Result<()> { ... }
    fn on_historical_quotes(&mut self, quotes: &[QuoteTick]) -> Result<()> { ... }
    fn on_historical_trades(&mut self, trades: &[TradeTick]) -> Result<()> { ... }
    fn on_historical_bars(&mut self, bars: &[Bar]) -> Result<()> { ... }
    fn on_historical_mark_prices(
        &mut self,
        mark_prices: &[MarkPriceUpdate],
    ) -> Result<()> { ... }
    fn on_historical_index_prices(
        &mut self,
        index_prices: &[IndexPriceUpdate],
    ) -> Result<()> { ... }
    fn handle_time_event(&mut self, event: &TimeEvent) { ... }
    fn handle_data(&mut self, data: &dyn Any) { ... }
    fn handle_signal(&mut self, signal: &Signal) { ... }
    fn handle_instrument(&mut self, instrument: &InstrumentAny) { ... }
    fn handle_book_deltas(&mut self, deltas: &OrderBookDeltas) { ... }
    fn handle_book(&mut self, book: &OrderBook) { ... }
    fn handle_quote(&mut self, quote: &QuoteTick) { ... }
    fn handle_trade(&mut self, trade: &TradeTick) { ... }
    fn handle_bar(&mut self, bar: &Bar) { ... }
    fn handle_mark_price(&mut self, mark_price: &MarkPriceUpdate) { ... }
    fn handle_index_price(&mut self, index_price: &IndexPriceUpdate) { ... }
    fn handle_instrument_status(&mut self, status: &InstrumentStatus) { ... }
    fn handle_instrument_close(&mut self, close: &InstrumentClose) { ... }
    fn handle_block(&mut self, block: &Block) { ... }
    fn handle_pool(&mut self, pool: &Pool) { ... }
    fn handle_pool_swap(&mut self, swap: &PoolSwap) { ... }
    fn handle_pool_liquidity_update(&mut self, update: &PoolLiquidityUpdate) { ... }
    fn handle_historical_data(&mut self, data: &dyn Any) { ... }
    fn handle_data_response(&mut self, resp: &CustomDataResponse) { ... }
    fn handle_instrument_response(&mut self, resp: &InstrumentResponse) { ... }
    fn handle_instruments_response(&mut self, resp: &InstrumentsResponse) { ... }
    fn handle_book_response(&mut self, resp: &BookResponse) { ... }
    fn handle_quotes_response(&mut self, resp: &QuotesResponse) { ... }
    fn handle_trades_response(&mut self, resp: &TradesResponse) { ... }
    fn handle_bars_response(&mut self, resp: &BarsResponse) { ... }
    fn subscribe_data(
        &mut self,
        data_type: DataType,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_quotes(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_instruments(
        &mut self,
        venue: Venue,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_instrument(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_book_deltas(
        &mut self,
        instrument_id: InstrumentId,
        book_type: BookType,
        depth: Option<NonZeroUsize>,
        client_id: Option<ClientId>,
        managed: bool,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_book_at_interval(
        &mut self,
        instrument_id: InstrumentId,
        book_type: BookType,
        depth: Option<NonZeroUsize>,
        interval_ms: NonZeroUsize,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_trades(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_bars(
        &mut self,
        bar_type: BarType,
        client_id: Option<ClientId>,
        await_partial: bool,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_mark_prices(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_index_prices(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_instrument_status(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_instrument_close(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_blocks(
        &mut self,
        chain: Blockchain,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_pool(
        &mut self,
        address: Address,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_pool_swaps(
        &mut self,
        address: Address,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn subscribe_pool_liquidity_updates(
        &mut self,
        address: Address,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_data(
        &mut self,
        data_type: DataType,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_instruments(
        &mut self,
        venue: Venue,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_instrument(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_book_deltas(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_book_at_interval(
        &mut self,
        instrument_id: InstrumentId,
        interval_ms: NonZeroUsize,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_quotes(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_trades(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_bars(
        &mut self,
        bar_type: BarType,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_mark_prices(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_index_prices(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_instrument_status(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_instrument_close(
        &mut self,
        instrument_id: InstrumentId,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_blocks(
        &mut self,
        chain: Blockchain,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_pool(
        &mut self,
        address: Address,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_pool_swaps(
        &mut self,
        address: Address,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn unsubscribe_pool_liquidity_updates(
        &mut self,
        address: Address,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    )
       where Self: 'static + Debug + Sized { ... }
    fn request_data(
        &mut self,
        data_type: DataType,
        client_id: ClientId,
        start: Option<DateTime<Utc>>,
        end: Option<DateTime<Utc>>,
        limit: Option<NonZeroUsize>,
        params: Option<IndexMap<String, String>>,
    ) -> Result<UUID4>
       where Self: 'static + Debug + Sized { ... }
    fn request_instrument(
        &mut self,
        instrument_id: InstrumentId,
        start: Option<DateTime<Utc>>,
        end: Option<DateTime<Utc>>,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    ) -> Result<UUID4>
       where Self: 'static + Debug + Sized { ... }
    fn request_instruments(
        &mut self,
        venue: Option<Venue>,
        start: Option<DateTime<Utc>>,
        end: Option<DateTime<Utc>>,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    ) -> Result<UUID4>
       where Self: 'static + Debug + Sized { ... }
    fn request_book_snapshot(
        &mut self,
        instrument_id: InstrumentId,
        depth: Option<NonZeroUsize>,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    ) -> Result<UUID4>
       where Self: 'static + Debug + Sized { ... }
    fn request_quotes(
        &mut self,
        instrument_id: InstrumentId,
        start: Option<DateTime<Utc>>,
        end: Option<DateTime<Utc>>,
        limit: Option<NonZeroUsize>,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    ) -> Result<UUID4>
       where Self: 'static + Debug + Sized { ... }
    fn request_trades(
        &mut self,
        instrument_id: InstrumentId,
        start: Option<DateTime<Utc>>,
        end: Option<DateTime<Utc>>,
        limit: Option<NonZeroUsize>,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    ) -> Result<UUID4>
       where Self: 'static + Debug + Sized { ... }
    fn request_bars(
        &mut self,
        bar_type: BarType,
        start: Option<DateTime<Utc>>,
        end: Option<DateTime<Utc>>,
        limit: Option<NonZeroUsize>,
        client_id: Option<ClientId>,
        params: Option<IndexMap<String, String>>,
    ) -> Result<UUID4>
       where Self: 'static + Debug + Sized { ... }

}

Provided Methods

fn on_save(&self) -> Result<IndexMap<String, Vec<u8>>>

Actions to be performed when the actor state is saved.

Errors

Returns an error if saving the actor state fails.

fn on_load(&mut self, state: IndexMap<String, Vec<u8>>) -> Result<()>

Actions to be performed when the actor state is loaded.

Errors

Returns an error if loading the actor state fails.

fn on_start(&mut self) -> Result<()>

Actions to be performed on start.

Errors

Returns an error if starting the actor fails.

fn on_stop(&mut self) -> Result<()>

Actions to be performed on stop.

Errors

Returns an error if stopping the actor fails.

fn on_resume(&mut self) -> Result<()>

Actions to be performed on resume.

Errors

Returns an error if resuming the actor fails.

fn on_reset(&mut self) -> Result<()>

Actions to be performed on reset.

Errors

Returns an error if resetting the actor fails.

fn on_dispose(&mut self) -> Result<()>

Actions to be performed on dispose.

Errors

Returns an error if disposing the actor fails.

fn on_degrade(&mut self) -> Result<()>

Actions to be performed on degrade.

Errors

Returns an error if degrading the actor fails.

fn on_fault(&mut self) -> Result<()>

Actions to be performed on fault.

Errors

Returns an error if faulting the actor fails.

fn on_time_event(&mut self, event: &TimeEvent) -> Result<()>

Actions to be performed when receiving a time event.

Errors

Returns an error if handling the time event fails.

fn on_data(&mut self, data: &dyn Any) -> Result<()>

Actions to be performed when receiving custom data.

Errors

Returns an error if handling the data fails.

fn on_signal(&mut self, signal: &Signal) -> Result<()>

Actions to be performed when receiving a signal.

Errors

Returns an error if handling the signal fails.

fn on_instrument(&mut self, instrument: &InstrumentAny) -> Result<()>

Actions to be performed when receiving an instrument.

Errors

Returns an error if handling the instrument fails.

fn on_book_deltas(&mut self, deltas: &OrderBookDeltas) -> Result<()>

Actions to be performed when receiving order book deltas.

Errors

Returns an error if handling the book deltas fails.

fn on_book(&mut self, order_book: &OrderBook) -> Result<()>

Actions to be performed when receiving an order book.

Errors

Returns an error if handling the book fails.

fn on_quote(&mut self, quote: &QuoteTick) -> Result<()>

Actions to be performed when receiving a quote.

Errors

Returns an error if handling the quote fails.

fn on_trade(&mut self, tick: &TradeTick) -> Result<()>

Actions to be performed when receiving a trade.

Errors

Returns an error if handling the trade fails.

fn on_bar(&mut self, bar: &Bar) -> Result<()>

Actions to be performed when receiving a bar.

Errors

Returns an error if handling the bar fails.

fn on_mark_price(&mut self, mark_price: &MarkPriceUpdate) -> Result<()>

Actions to be performed when receiving a mark price update.

Errors

Returns an error if handling the mark price update fails.

fn on_index_price(&mut self, index_price: &IndexPriceUpdate) -> Result<()>

Actions to be performed when receiving an index price update.

Errors

Returns an error if handling the index price update fails.

fn on_instrument_status(&mut self, data: &InstrumentStatus) -> Result<()>

Actions to be performed when receiving an instrument status update.

Errors

Returns an error if handling the instrument status update fails.

fn on_instrument_close(&mut self, update: &InstrumentClose) -> Result<()>

Actions to be performed when receiving an instrument close update.

Errors

Returns an error if handling the instrument close update fails.

fn on_block(&mut self, block: &Block) -> Result<()>

Actions to be performed when receiving a block.

Errors

Returns an error if handling the block fails.

fn on_pool(&mut self, pool: &Pool) -> Result<()>

Actions to be performed when receiving a pool.

Errors

Returns an error if handling the pool fails.

fn on_pool_swap(&mut self, swap: &PoolSwap) -> Result<()>

Actions to be performed when receiving a pool swap.

Errors

Returns an error if handling the pool swap fails.

fn on_pool_liquidity_update( &mut self, update: &PoolLiquidityUpdate, ) -> Result<()>

Actions to be performed when receiving a pool liquidity update.

Errors

Returns an error if handling the pool liquidity update fails.

fn on_historical_data(&mut self, data: &dyn Any) -> Result<()>

Actions to be performed when receiving historical data.

Errors

Returns an error if handling the historical data fails.

fn on_historical_quotes(&mut self, quotes: &[QuoteTick]) -> Result<()>

Actions to be performed when receiving historical quotes.

Errors

Returns an error if handling the historical quotes fails.

fn on_historical_trades(&mut self, trades: &[TradeTick]) -> Result<()>

Actions to be performed when receiving historical trades.

Errors

Returns an error if handling the historical trades fails.

fn on_historical_bars(&mut self, bars: &[Bar]) -> Result<()>

Actions to be performed when receiving historical bars.

Errors

Returns an error if handling the historical bars fails.

fn on_historical_mark_prices( &mut self, mark_prices: &[MarkPriceUpdate], ) -> Result<()>

Actions to be performed when receiving historical mark prices.

Errors

Returns an error if handling the historical mark prices fails.

fn on_historical_index_prices( &mut self, index_prices: &[IndexPriceUpdate], ) -> Result<()>

Actions to be performed when receiving historical index prices.

Errors

Returns an error if handling the historical index prices fails.

fn handle_time_event(&mut self, event: &TimeEvent)

Handles a received time event.

fn handle_data(&mut self, data: &dyn Any)

Handles a received custom data point.

fn handle_signal(&mut self, signal: &Signal)

Handles a received signal.

fn handle_instrument(&mut self, instrument: &InstrumentAny)

Handles a received instrument.

fn handle_book_deltas(&mut self, deltas: &OrderBookDeltas)

Handles received order book deltas.

fn handle_book(&mut self, book: &OrderBook)

Handles a received order book reference.

fn handle_quote(&mut self, quote: &QuoteTick)

Handles a received quote.

fn handle_trade(&mut self, trade: &TradeTick)

Handles a received trade.

fn handle_bar(&mut self, bar: &Bar)

Handles a receiving bar.

fn handle_mark_price(&mut self, mark_price: &MarkPriceUpdate)

Handles a received mark price update.

fn handle_index_price(&mut self, index_price: &IndexPriceUpdate)

Handles a received index price update.

fn handle_instrument_status(&mut self, status: &InstrumentStatus)

Handles a received instrument status.

fn handle_instrument_close(&mut self, close: &InstrumentClose)

Handles a received instrument close.

fn handle_block(&mut self, block: &Block)

Handles a received block.

fn handle_pool(&mut self, pool: &Pool)

Handles a received pool definition update.

fn handle_pool_swap(&mut self, swap: &PoolSwap)

Handles a received pool swap.

fn handle_pool_liquidity_update(&mut self, update: &PoolLiquidityUpdate)

Handles a received pool liquidity update.

fn handle_historical_data(&mut self, data: &dyn Any)

Handles received historical data.

fn handle_data_response(&mut self, resp: &CustomDataResponse)

Handles a data response.

fn handle_instrument_response(&mut self, resp: &InstrumentResponse)

Handles an instrument response.

fn handle_instruments_response(&mut self, resp: &InstrumentsResponse)

Handles an instruments response.

fn handle_book_response(&mut self, resp: &BookResponse)

Handles a book response.

fn handle_quotes_response(&mut self, resp: &QuotesResponse)

Handles a quotes response.

fn handle_trades_response(&mut self, resp: &TradesResponse)

Handles a trades response.

fn handle_bars_response(&mut self, resp: &BarsResponse)

Handles a bars response.

fn subscribe_data( &mut self, data_type: DataType, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming data_type data.

fn subscribe_quotes( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [QuoteTick] data for the instrument_id.

fn subscribe_instruments( &mut self, venue: Venue, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [InstrumentAny] data for the venue.

fn subscribe_instrument( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [InstrumentAny] data for the instrument_id.

fn subscribe_book_deltas( &mut self, instrument_id: InstrumentId, book_type: BookType, depth: Option<NonZeroUsize>, client_id: Option<ClientId>, managed: bool, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [OrderBookDeltas] data for the instrument_id.

fn subscribe_book_at_interval( &mut self, instrument_id: InstrumentId, book_type: BookType, depth: Option<NonZeroUsize>, interval_ms: NonZeroUsize, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to [OrderBook] snapshots at a specified interval for the instrument_id.

fn subscribe_trades( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [TradeTick] data for the instrument_id.

fn subscribe_bars( &mut self, bar_type: BarType, client_id: Option<ClientId>, await_partial: bool, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [Bar] data for the bar_type.

fn subscribe_mark_prices( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [MarkPriceUpdate] data for the instrument_id.

fn subscribe_index_prices( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [IndexPriceUpdate] data for the instrument_id.

fn subscribe_instrument_status( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [InstrumentStatus] data for the instrument_id.

fn subscribe_instrument_close( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [InstrumentClose] data for the instrument_id.

fn subscribe_blocks( &mut self, chain: Blockchain, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [Block] data for the chain.

fn subscribe_pool( &mut self, address: Address, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [Pool] definition updates for the AMM pool at the address.

fn subscribe_pool_swaps( &mut self, address: Address, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [PoolSwap] data for the address.

fn subscribe_pool_liquidity_updates( &mut self, address: Address, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Subscribe to streaming [PoolLiquidityUpdate] data for the address.

fn unsubscribe_data( &mut self, data_type: DataType, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming data_type data.

fn unsubscribe_instruments( &mut self, venue: Venue, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [InstrumentAny] data for the venue.

fn unsubscribe_instrument( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [InstrumentAny] data for the instrument_id.

fn unsubscribe_book_deltas( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [OrderBookDeltas] data for the instrument_id.

fn unsubscribe_book_at_interval( &mut self, instrument_id: InstrumentId, interval_ms: NonZeroUsize, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from [OrderBook] snapshots at a specified interval for the instrument_id.

fn unsubscribe_quotes( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [QuoteTick] data for the instrument_id.

fn unsubscribe_trades( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [TradeTick] data for the instrument_id.

fn unsubscribe_bars( &mut self, bar_type: BarType, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [Bar] data for the bar_type.

fn unsubscribe_mark_prices( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [MarkPriceUpdate] data for the instrument_id.

fn unsubscribe_index_prices( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [IndexPriceUpdate] data for the instrument_id.

fn unsubscribe_instrument_status( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [InstrumentStatus] data for the instrument_id.

fn unsubscribe_instrument_close( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [InstrumentClose] data for the instrument_id.

fn unsubscribe_blocks( &mut self, chain: Blockchain, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [Block] data for the chain.

fn unsubscribe_pool( &mut self, address: Address, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [Pool] definition updates for the AMM pool at the address.

fn unsubscribe_pool_swaps( &mut self, address: Address, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [PoolSwap] data for the address.

fn unsubscribe_pool_liquidity_updates( &mut self, address: Address, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )
where Self: 'static + Debug + Sized,

Unsubscribe from streaming [PoolLiquidityUpdate] data for the address.

fn request_data( &mut self, data_type: DataType, client_id: ClientId, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, limit: Option<NonZeroUsize>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>
where Self: 'static + Debug + Sized,

Request historical custom data of the given data_type.

Errors

Returns an error if input parameters are invalid.

fn request_instrument( &mut self, instrument_id: InstrumentId, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>
where Self: 'static + Debug + Sized,

Request historical InstrumentResponse data for the given instrument_id.

Errors

Returns an error if input parameters are invalid.

fn request_instruments( &mut self, venue: Option<Venue>, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>
where Self: 'static + Debug + Sized,

Request historical InstrumentsResponse definitions for the optional venue.

Errors

Returns an error if input parameters are invalid.

fn request_book_snapshot( &mut self, instrument_id: InstrumentId, depth: Option<NonZeroUsize>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>
where Self: 'static + Debug + Sized,

Request an [OrderBook] snapshot for the given instrument_id.

Errors

Returns an error if input parameters are invalid.

fn request_quotes( &mut self, instrument_id: InstrumentId, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, limit: Option<NonZeroUsize>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>
where Self: 'static + Debug + Sized,

Request historical [QuoteTick] data for the given instrument_id.

Errors

Returns an error if input parameters are invalid.

fn request_trades( &mut self, instrument_id: InstrumentId, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, limit: Option<NonZeroUsize>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>
where Self: 'static + Debug + Sized,

Request historical [TradeTick] data for the given instrument_id.

Errors

Returns an error if input parameters are invalid.

fn request_bars( &mut self, bar_type: BarType, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, limit: Option<NonZeroUsize>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>
where Self: 'static + Debug + Sized,

Request historical [Bar] data for the given bar_type.

Errors

Returns an error if input parameters are invalid.