nautilus_common::clock

Struct TestClock

pub struct TestClock { /* private fields */ }

A static test clock.

Stores the current timestamp internally which can be advanced.

Implementations

impl TestClock

pub fn new() -> Self

Creates a new TestClock instance.

pub const fn get_timers(&self) -> &BTreeMap<Ustr, TestTimer>

Returns a reference to the internal timers for the clock.

pub fn advance_time( &mut self, to_time_ns: UnixNanos, set_time: bool, ) -> Vec<TimeEvent>

Advances the internal clock to the specified to_time_ns and optionally sets the clock to that time.

This function ensures that the clock behaves in a non-decreasing manner. If set_time is true, the internal clock will be updated to the value of to_time_ns. Otherwise, the clock will advance without explicitly setting the time.

The method processes active timers, advancing them to to_time_ns, and collects any TimeEvent objects that are triggered as a result. Only timers that are not expired are processed.

Panics

Panics if to_time_ns is less than the current internal clock time.

pub fn advance_to_time_on_heap(&mut self, to_time_ns: UnixNanos)

Advances the internal clock to the specified to_time_ns and optionally sets the clock to that time.

Pushes the TimeEvents on the heap to ensure ordering

Note: set_time is not used but present to keep backward compatible api call

Panics

Panics if to_time_ns is less than the current internal clock time.

pub fn match_handlers(&self, events: Vec<TimeEvent>) -> Vec<TimeEventHandlerV2>

Matches TimeEvent objects with their corresponding event handlers.

This function takes an events vector of TimeEvent objects, assumes they are already sorted by their ts_event, and matches them with the appropriate callback handler from the internal registry of callbacks. If no specific callback is found for an event, the default callback is used.

Panics

Panics if the default callback is not set for the clock when matching handlers.