Trait nautilus_common::clock::Clock

pub trait Clock {

    // Required methods
    fn new() -> Self;
    fn timestamp(&mut self) -> f64;
    fn timestamp_ms(&mut self) -> u64;
    fn timestamp_us(&mut self) -> u64;
    fn timestamp_ns(&mut self) -> u64;
    fn timer_names(&self) -> Vec<&str>;
    fn timer_count(&self) -> usize;
    fn register_default_handler(&mut self, callback: Box<dyn Fn(TimeEvent)>);
    fn register_default_handler_py(&mut self, callback_py: PyObject);
    fn set_time_alert_ns_py(
        &mut self,
        name: String,
        alert_time_ns: UnixNanos,
        callback_py: Option<PyObject>
    );
    fn set_timer_ns_py(
        &mut self,
        name: String,
        interval_ns: u64,
        start_time_ns: UnixNanos,
        stop_time_ns: Option<UnixNanos>,
        callback_py: Option<PyObject>
    );
    fn next_time_ns(&mut self, name: &str) -> UnixNanos;
    fn cancel_timer(&mut self, name: &str);
    fn cancel_timers(&mut self);

}

Represents a type of clock.

Notes

An active timer is one which has not expired (timer.is_expired == False).

Required Methods

fn new() -> Self

Return a new Clock.

fn timestamp(&mut self) -> f64

Return the current UNIX time in seconds.

fn timestamp_ms(&mut self) -> u64

Return the current UNIX time in milliseconds (ms).

fn timestamp_us(&mut self) -> u64

Return the current UNIX time in microseconds (us).

fn timestamp_ns(&mut self) -> u64

Return the current UNIX time in nanoseconds (ns).

fn timer_names(&self) -> Vec<&str>

Return the names of active timers in the clock.

fn timer_count(&self) -> usize

Return the count of active timers in the clock.

fn register_default_handler(&mut self, callback: Box<dyn Fn(TimeEvent)>)

Register a default event handler for the clock. If a [Timer] does not have an event handler, then this handler is used.

fn register_default_handler_py(&mut self, callback_py: PyObject)

fn set_time_alert_ns_py( &mut self, name: String, alert_time_ns: UnixNanos, callback_py: Option<PyObject> )

Set a [Timer] to alert at a particular time. Optional callback gets used to handle generated events.

fn set_timer_ns_py( &mut self, name: String, interval_ns: u64, start_time_ns: UnixNanos, stop_time_ns: Option<UnixNanos>, callback_py: Option<PyObject> )

Set a [Timer] to start alerting at every interval between start and stop time. Optional callback gets used to handle generated event.

fn next_time_ns(&mut self, name: &str) -> UnixNanos

fn cancel_timer(&mut self, name: &str)

fn cancel_timers(&mut self)