nautilus_common::component

Trait Component

pub trait Component: Debug + Any {
    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn id(&self) -> ComponentId;
    fn state(&self) -> ComponentState;
    fn is_running(&self) -> bool;
    fn is_stopped(&self) -> bool;
    fn is_disposed(&self) -> bool;
    fn register(
        &mut self,
        trader_id: TraderId,
        clock: Rc<RefCell<dyn Clock>>,
        cache: Rc<RefCell<Cache>>,
    ) -> Result<()>;
    fn start(&mut self) -> Result<()>;
    fn stop(&mut self) -> Result<()>;
    fn reset(&mut self) -> Result<()>;
    fn dispose(&mut self) -> Result<()>;
    fn handle_event(&mut self, event: TimeEvent);
}

Common trait for components.

Required Methods

fn as_any(&self) -> &dyn Any

Returns a reference to the component as Any for downcasting.

fn id(&self) -> ComponentId

Returns the unique identifier for this component.

fn state(&self) -> ComponentState

Returns the current state of the component.

fn is_running(&self) -> bool

Returns whether the component is currently running.

fn is_stopped(&self) -> bool

Returns whether the component is stopped.

fn is_disposed(&self) -> bool

Returns whether the component has been disposed.

fn register( &mut self, trader_id: TraderId, clock: Rc<RefCell<dyn Clock>>, cache: Rc<RefCell<Cache>>, ) -> Result<()>

Registers the component with a system.

Errors

Returns an error if the component fails to register.

fn start(&mut self) -> Result<()>

Starts the component.

Errors

Returns an error if the component fails to start.

fn stop(&mut self) -> Result<()>

Stops the component.

Errors

Returns an error if the component fails to stop.

fn reset(&mut self) -> Result<()>

Resets the component to its initial state.

Errors

Returns an error if the component fails to reset.

fn dispose(&mut self) -> Result<()>

Disposes of the component, releasing any resources.

Errors

Returns an error if the component fails to dispose.

fn handle_event(&mut self, event: TimeEvent)

Handles a timer event (TBD).