nautilus_common

Module enums

Enumerations for common components.

Structs

ComponentStateIter

An iterator over the variants of ComponentState

ComponentTriggerIter

An iterator over the variants of ComponentTrigger

EnvironmentIter

An iterator over the variants of Environment

LogColorIter

An iterator over the variants of LogColor

LogLevelIter

An iterator over the variants of LogLevel

SerializationEncodingIter

An iterator over the variants of SerializationEncoding

Enums

ComponentState

The state of a component within the system.

ComponentTrigger

A trigger condition for a component within the system.

Environment

Represents the environment context for a Posei system.

LogColor

The log color for log messages.

LogFormat

An ANSI log line format specifier. This is used for formatting log messages with ANSI escape codes.

LogLevel

The log level for log messages.

SerializationEncoding

The serialization encoding.