Enum nautilus_common::enums::ComponentTrigger

#[repr(C)]
pub enum ComponentTrigger {

    Initialize,
    Start,
    StartCompleted,
    Stop,
    StopCompleted,
    Resume,
    ResumeCompleted,
    Reset,
    ResetCompleted,
    Dispose,
    DisposeCompleted,
    Degrade,
    DegradeCompleted,
    Fault,
    FaultCompleted,

}

A trigger condition for a component within the system.

Variants

Initialize

A trigger for the component to initialize.

Start

A trigger for the component to start.

StartCompleted

A trigger when the component has successfully started.

Stop

A trigger for the component to stop.

StopCompleted

A trigger when the component has successfully stopped.

Resume

A trigger for the component to resume (after being stopped).

ResumeCompleted

A trigger when the component has successfully resumed.

Reset

A trigger for the component to reset.

ResetCompleted

A trigger when the component has successfully reset.

Dispose

A trigger for the component to dispose and release resources.

DisposeCompleted

A trigger when the component has successfully disposed.

Degrade

A trigger for the component to degrade.

DegradeCompleted

A trigger when the component has successfully degraded.

Fault

A trigger for the component to fault.

FaultCompleted

A trigger when the component has successfully faulted.

Implementations

impl ComponentTrigger

pub const fn from_repr(discriminant: usize) -> Option<ComponentTrigger>

Try to create Self from the raw representation