Enum nautilus_common::enums::LogFormat

#[repr(C)]
pub enum LogFormat {
    Header = 0,
    Endc = 1,
    Bold = 2,
    Underline = 3,
}

An ANSI log line format specifier. This is used for formatting log messages with ANSI escape codes.

Variants

Header = 0

Header log format. This ANSI escape code is used for magenta text color, often used for headers or titles in the log output.

Endc = 1

Endc log format. This ANSI escape code is used to reset all format attributes to their defaults. It should be used after applying other formats.

Bold = 2

Bold log format. This ANSI escape code is used to make the text bold in the log output.

Underline = 3

Underline log format. This ANSI escape code is used to underline the text in the log output.

Implementations

impl LogFormat

pub const fn from_repr(discriminant: usize) -> Option<LogFormat>

Try to create Self from the raw representation