nautilus_common::ffi::clock

Struct LiveClock_API

#[repr(C)]
pub struct LiveClock_API(/* private fields */);

C compatible Foreign Function Interface (FFI) for an underlying LiveClock.

This struct wraps LiveClock in a way that makes it compatible with C function calls, enabling interaction with LiveClock in a C environment.

It implements the Deref and DerefMut traits, allowing instances of LiveClock_API to be dereferenced to LiveClock, providing access to LiveClock’s methods without having to manually access the underlying LiveClock instance. This includes both mutable and immutable access.