Struct nautilus_common::ffi::msgbus::MessageBus_API

#[repr(C)]
pub struct MessageBus_API(/* private fields */);

Provides a C compatible Foreign Function Interface (FFI) for an underlying MessageBus.

This struct wraps MessageBus in a way that makes it compatible with C function calls, enabling interaction with MessageBus in a C environment.

It implements the Deref trait, allowing instances of MessageBus_API to be dereferenced to MessageBus, providing access to TestClock’s methods without having to manually access the underlying MessageBus instance.