nautilus_common::logging::logger

Struct LoggerConfig

pub struct LoggerConfig {
    pub stdout_level: LevelFilter,
    pub fileout_level: LevelFilter,
    pub is_colored: bool,
    pub print_config: bool,
    /* private fields */
}

Fields
stdout_level: LevelFilter

Maximum log level to write to stdout.

fileout_level: LevelFilter

Maximum log level to write to file (disabled is Off).

is_colored: bool

If logger is using ANSI color codes.

print_config: bool

If the configuration should be printed to stdout at initialization.

Implementations

impl LoggerConfig

pub const fn new( stdout_level: LevelFilter, fileout_level: LevelFilter, component_level: HashMap<Ustr, LevelFilter>, is_colored: bool, print_config: bool, ) -> Self

Creates a new LoggerConfig instance.

pub fn from_spec(spec: &str) -> Result<Self>

Errors

Returns an error if the spec string is invalid.

pub fn from_env() -> Result<Self>

Retrieves the logger configuration from the “NAUTILUS_LOG” environment variable.

Errors

Returns an error if the variable is unset or invalid.

impl LoggerConfig

pub fn py_from_spec(spec: String) -> PyResult<Self>

Creates a LoggerConfig from a spec string.

Errors

Returns a Python exception if the spec string is invalid.