Struct nautilus_common::logging::Logger

pub struct Logger {
    pub trader_id: TraderId,
    pub machine_id: String,
    pub instance_id: UUID4,
    pub level_stdout: LogLevel,
    pub level_file: Option<LogLevel>,
    pub is_bypassed: bool,
    /* private fields */
}

Provides a high-performance logger utilizing a MPSC channel under the hood.

A separate thead is spawned at initialization which receives LogEvent structs over the channel.

Fields
trader_id: TraderId

The trader ID for the logger.

machine_id: String

The machine ID for the logger.

instance_id: UUID4

The instance ID for the logger.

level_stdout: LogLevel

The minimum log level to write to stdout.

level_file: Option<LogLevel>

The minimum log level to write to a log file.

is_bypassed: bool

If logging is bypassed.

Implementations

impl Logger

pub fn new( trader_id: TraderId, machine_id: String, instance_id: UUID4, level_stdout: LogLevel, level_file: Option<LogLevel>, directory: Option<String>, file_name: Option<String>, file_format: Option<String>, component_levels: Option<HashMap<String, Value>>, is_bypassed: bool ) -> Self

pub fn send( &mut self, timestamp: u64, level: LogLevel, color: LogColor, component: String, message: String )

pub fn debug( &mut self, timestamp: u64, color: LogColor, component: String, message: String )

pub fn info( &mut self, timestamp: u64, color: LogColor, component: String, message: String )

pub fn warn( &mut self, timestamp: u64, color: LogColor, component: String, message: String )

pub fn error( &mut self, timestamp: u64, color: LogColor, component: String, message: String )

pub fn critical( &mut self, timestamp: u64, color: LogColor, component: String, message: String )