Struct nautilus_common::msgbus::MessageBus

pub struct MessageBus {
    pub trader_id: TraderId,
    pub instance_id: UUID4,
    pub name: String,
    pub sent_count: u64,
    pub req_count: u64,
    pub res_count: u64,
    pub pub_count: u64,
    pub has_backing: bool,
    /* private fields */
}

Provides a generic message bus to facilitate various messaging patterns.

The bus provides both a producer and consumer API for Pub/Sub, Req/Rep, as well as direct point-to-point messaging to registered endpoints.

Pub/Sub wildcard patterns for hierarchical topics are possible:

* asterisk represents one or more characters in a pattern.

? question mark represents a single character in a pattern.

Given a topic and pattern potentially containing wildcard characters, i.e. * and ?, where ? can match any single character in the topic, and * can match any number of characters including zero characters.

The asterisk in a wildcard matches any character zero or more times. For example, comp* matches anything beginning with comp which means comp, complete, and computer are all matched.

A question mark matches a single character once. For example, c?mp matches camp and comp. The question mark can also be used more than once. For example, c??p would match both of the above examples and coop.

Fields
trader_id: TraderId

The trader ID associated with the message bus.

instance_id: UUID4

The instance ID associated with the message bus.

name: String

The name for the message bus.

sent_count: u64req_count: u64res_count: u64pub_count: u64

The count of messages published by the bus.

has_backing: bool

If the message bus is backed by a database.

Implementations

impl MessageBus

pub fn new( trader_id: TraderId, instance_id: UUID4, name: Option<String>, config: Option<HashMap<String, Value>> ) -> Self

Initializes a new instance of the MessageBus.

pub fn endpoints(&self) -> Vec<&str>

Returns the registered endpoint addresses.

pub fn topics(&self) -> Vec<&str>

Returns the topics for active subscriptions.

pub fn correlation_ids(&self) -> Vec<&UUID4>

Returns the active correlation IDs.

pub fn has_subscribers(&self, pattern: &str) -> bool

Returns whether there are subscribers for the given pattern.

pub fn subscriptions(&self) -> Vec<&Subscription>

Returns whether there are subscribers for the given pattern.

pub fn subscription_handler_ids(&self) -> Vec<&str>

Returns whether there are subscribers for the given pattern.

pub fn is_registered(&self, endpoint: &str) -> bool

Returns whether there are subscribers for the given pattern.

pub fn is_subscribed(&self, topic: &str, handler: MessageHandler) -> bool

Returns whether there are subscribers for the given pattern.

pub fn is_pending_response(&self, request_id: &UUID4) -> bool

Returns whether there is a pending request for the given request_id.

pub fn register(&mut self, endpoint: &str, handler: MessageHandler)

Registers the given handler for the endpoint address.

pub fn deregister(&mut self, endpoint: &str)

Deregisters the given handler for the endpoint address.

pub fn subscribe( &mut self, topic: &str, handler: MessageHandler, priority: Option<u8> )

Subscribes the given handler to the topic.

pub fn unsubscribe(&mut self, topic: &str, handler: MessageHandler)

Unsubscribes the given handler from the topic.

pub fn get_endpoint(&self, endpoint: &Ustr) -> Option<&MessageHandler>

Returns the handler for the given endpoint.

pub fn request_handler( &mut self, endpoint: &Ustr, request_id: UUID4, response_handler: MessageHandler ) -> Option<&MessageHandler>

Returns the handler for the request endpoint and adds the request ID to the internal correlation index to match with the expected response.

pub fn correlation_id_handler( &mut self, correlation_id: &UUID4 ) -> Option<&MessageHandler>

Returns the handler for the matching correlation ID (if found).

pub fn response_handler( &mut self, correlation_id: &UUID4 ) -> Option<MessageHandler>

Returns the handler for the matching response endpoint based on the internal correlation index.

pub fn matching_subscriptions<'a>( &'a self, pattern: &'a Ustr ) -> Vec<&'a Subscription>

pub fn publish_external(&self, topic: String, payload: Vec<u8>)