nautilus_common::python::actor

Struct PyDataActor

pub struct PyDataActor { /* private fields */ }

Provides a generic DataActor.

Implementations

impl PyDataActor

pub fn register( &mut self, trader_id: TraderId, cache: Rc<RefCell<Cache>>, clock: Rc<RefCell<dyn Clock>>, ) -> PyResult<()>

TODO: WIP This method should be called to properly initialize the actor with cache, clock and other components.

Errors

Returns an error if already registered.