Function nautilus_common::python::logging::py_init_tracing

pub fn py_init_tracing()

Initialize tracing.

Tracing is meant to be used to trace/debug async Rust code. It can be configured to filter modules and write up to a specific level only using by passing a configuration using the RUST_LOG environment variable.

Safety

Should only be called once during an applications run, ideally at the beginning of the run.