nautilus_common::python

Module msgbus

Functions

py_msgbus_deregister

Deregisters the given handler for the endpoint address.

py_msgbus_is_registered

Returns whether there are subscribers for the given pattern.

py_msgbus_is_subscribed

Returns whether there are subscribers for the given pattern.

py_msgbus_publish

Publishes the message to the topic.

py_msgbus_register

Registers the given handler for the endpoint address.

py_msgbus_send

Sends the message to the endpoint.

py_msgbus_subscribe

Subscribes the given handler to the topic.

py_msgbus_unsubscribe

Unsubscribes the given handler from the topic.