Crate nautilus_common

NautilusTrader is an open-source, high-performance, production-grade algorithmic trading platform, providing quantitative traders with the ability to backtest portfolios of automated trading strategies on historical data with an event-driven engine, and also deploy those same strategies live, with no code changes.

Feature flags

This crate provides feature flags to control source code inclusion during compilation, depending on the intended use case, i.e. whether to provide Python bindings for the main nautilus_trader Python package, or as part of a Rust only build.

ffi: Enables the C foreign function interface (FFI) from cbindgen

python: Enables Python bindings from pyo3

stubs: Enables type stubs for use in testing scenarios

Modules

cache

A common in-memory Cache for market and execution related data.

clock

Provides real-time and static test Clock implementations.

enums

Defines common enumerations.

factories

Provides factories for constructing domain objects such as orders.

ffi

Provides a C foreign function interface (FFI) from cbindgen.

generators

Provides generation of identifiers such as ClientOrderId and PositionId.

handlers

Provides common message handlers.

interface

Interface traits to faciliate a ports and adapters style architecture.

logging

The logging framework for Nautilus systems.

msgbus

A common in-memory MessageBus for loosely coupled message passing patterns.

python

Python bindings from pyo3.

runtime

The centralized Tokio runtime for a running Nautilus system.

stubs

Type stubs to facilitate testing.

testing

Common test related helper functions.

timer

Provides real-time and test timers for use with Clock implementations.

xrate

Provides exchange rate calculations between currencies.