nautilus_common::actor::data_actor

Struct DataActorCore

pub struct DataActorCore {
    pub actor_id: ActorId,
    pub config: DataActorConfig,
    pub clock: Rc<RefCell<dyn Clock>>,
    pub cache: Rc<RefCell<Cache>>,
    /* private fields */
}

Core functionality for all actors.

Fields
actor_id: ActorId

The actor identifier.

config: DataActorConfig

The actors configuration.

clock: Rc<RefCell<dyn Clock>>

The actors clock.

cache: Rc<RefCell<Cache>>

The cache for the actor.

Implementations

impl DataActorCore

pub fn new( config: DataActorConfig, cache: Rc<RefCell<Cache>>, clock: Rc<RefCell<dyn Clock>>, ) -> Self

Creates a new DataActorCore instance.

pub fn initialize(&mut self) -> Result<()>

Initializes the actor.

Errors

Returns an error if the initialization state transition fails.

pub fn trader_id(&self) -> Option<TraderId>

Returns the trader ID this actor is registered to.

pub fn state(&self) -> ComponentState

pub fn register_warning_event(&mut self, event_type: &str)

Register an event type for warning log levels.

pub fn deregister_warning_event(&mut self, event_type: &str)

Deregister an event type from warning log levels.

pub fn start(&mut self) -> Result<()>

Starts the actor.

Errors

Returns an error if starting the actor fails.

pub fn stop(&mut self) -> Result<()>

Stops the actor.

Errors

Returns an error if stopping the actor fails.

pub fn resume(&mut self) -> Result<()>

Resumes the actor.

Errors

Returns an error if resuming the actor fails.

pub fn reset(&mut self) -> Result<()>

Resets the actor.

Errors

Returns an error if resetting the actor fails.

pub fn dispose(&mut self) -> Result<()>

Disposes the actor.

Errors

Returns an error if disposing the actor fails.

pub fn degrade(&mut self) -> Result<()>

Degrades the actor.

Errors

Returns an error if degrading the actor fails.

pub fn fault(&mut self) -> Result<()>

Faults the actor.

Errors

Returns an error if faulting the actor fails.

pub fn shutdown_system(&self, reason: Option<String>)

Sends a shutdown command to the system with an optional reason.

Panics

Panics if the actor is not registered or has no trader ID.

pub fn subscribe_data<A: DataActor>( &mut self, data_type: DataType, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Subscribe to streaming data_type data.

pub fn subscribe_instruments<A: DataActor>( &mut self, venue: Venue, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Subscribe to streaming [InstrumentAny] data for the venue.

pub fn subscribe_instrument<A: DataActor>( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Subscribe to streaming [InstrumentAny] data for the instrument_id.

pub fn subscribe_book_deltas<A: DataActor>( &mut self, instrument_id: InstrumentId, book_type: BookType, depth: Option<NonZeroUsize>, client_id: Option<ClientId>, managed: bool, params: Option<IndexMap<String, String>>, )

Subscribe to streaming [OrderBookDeltas] data for the instrument_id.

Once subscribed, any matching order book deltas published on the message bus are forwarded to the on_book_deltas handler.

pub fn subscribe_book_at_interval<A: DataActor>( &mut self, instrument_id: InstrumentId, book_type: BookType, depth: Option<NonZeroUsize>, interval_ms: NonZeroUsize, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Subscribe to [OrderBook] snapshots at a specified interval for the instrument_id.

Once subscribed, any matching order book snapshots published on the message bus are forwarded to the on_book handler.

Warnings

Consider subscribing to order book deltas if you need intervals less than 100 milliseconds.

pub fn subscribe_quotes<A: DataActor>( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Subscribe to streaming [QuoteTick] data for the instrument_id.

pub fn subscribe_trades<A: DataActor>( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Subscribe to streaming [TradeTick] data for the instrument_id.

pub fn subscribe_bars<A: DataActor>( &mut self, bar_type: BarType, client_id: Option<ClientId>, await_partial: bool, params: Option<IndexMap<String, String>>, )

Subscribe to streaming [Bar] data for the bar_type.

Once subscribed, any matching bar data published on the message bus is forwarded to the on_bar handler.

pub fn subscribe_mark_prices<A: DataActor>( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Subscribe to streaming [MarkPriceUpdate] data for the instrument_id.

Once subscribed, any matching mark price updates published on the message bus are forwarded to the on_mark_price handler.

pub fn subscribe_index_prices<A: DataActor>( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Subscribe to streaming [IndexPriceUpdate] data for the instrument_id.

Once subscribed, any matching index price updates published on the message bus are forwarded to the on_index_price handler.

pub fn subscribe_instrument_status<A: DataActor>( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Subscribe to streaming [InstrumentStatus] data for the instrument_id.

Once subscribed, any matching bar data published on the message bus is forwarded to the on_bar handler.

pub fn subscribe_instrument_close<A: DataActor>( &mut self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Subscribe to streaming [InstrumentClose] data for the instrument_id.

Once subscribed, any matching instrument close data published on the message bus is forwarded to the on_instrument_close handler.

pub fn unsubscribe_data<A: DataActor>( &self, data_type: DataType, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from streaming data_type data.

pub fn unsubscribe_instruments<A: DataActor>( &self, venue: Venue, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from streaming [Instrument] data for the venue.

pub fn unsubscribe_instrument<A: DataActor>( &self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from streaming [Instrument] definitions for the instrument_id.

pub fn unsubscribe_book_deltas<A: DataActor>( &self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from streaming [OrderBookDeltas] for the instrument_id.

pub fn unsubscribe_book_at_interval<A: DataActor>( &mut self, instrument_id: InstrumentId, interval_ms: NonZeroUsize, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from [OrderBook] snapshots at a specified interval for the instrument_id.

The interval_ms must match a previously subscribed interval.

pub fn unsubscribe_quotes<A: DataActor>( &self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from streaming [QuoteTick] data for the instrument_id.

pub fn unsubscribe_trades<A: DataActor>( &self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from streaming [TradeTick] data for the instrument_id.

pub fn unsubscribe_bars<A: DataActor>( &mut self, bar_type: BarType, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from streaming [Bar] data for the bar_type.

pub fn unsubscribe_mark_prices<A: DataActor>( &self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from streaming [MarkPriceUpdate] data for the instrument_id.

pub fn unsubscribe_index_prices<A: DataActor>( &self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from streaming [IndexPriceUpdate] data for the instrument_id.

pub fn unsubscribe_instrument_status<A: DataActor>( &self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from streaming [InstrumentStatus] data for the instrument_id.

pub fn unsubscribe_instrument_close<A: DataActor>( &self, instrument_id: InstrumentId, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, )

Unsubscribe from streaming [InstrumentClose] data for the instrument_id.

pub fn request_data<A: DataActor>( &self, data_type: DataType, client_id: ClientId, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, limit: Option<NonZeroUsize>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>

Request historical custom data of the given data_type.

Returns a unique request ID to correlate subsequent CustomDataResponse.

Errors

Returns an error if the provided time range is invalid.

pub fn request_instrument<A: DataActor>( &self, instrument_id: InstrumentId, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>

Request historical InstrumentResponse data for the given instrument_id.

Returns a unique request ID to correlate subsequent InstrumentResponse.

Errors

Returns an error if the provided time range is invalid.

pub fn request_instruments<A: DataActor>( &self, venue: Option<Venue>, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>

Request historical InstrumentsResponse definitions for the optional venue.

Returns a unique request ID to correlate subsequent InstrumentsResponse.

Errors

Returns an error if the provided time range is invalid.

pub fn request_book_snapshot<A: DataActor>( &self, instrument_id: InstrumentId, depth: Option<NonZeroUsize>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>

Request an [OrderBook] snapshot for the given instrument_id.

Returns a unique request ID to correlate subsequent BookResponse.

Errors

This function never returns an error.

pub fn request_quotes<A: DataActor>( &self, instrument_id: InstrumentId, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, limit: Option<NonZeroUsize>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>

Request historical [QuoteTick] data for the given instrument_id.

Returns a unique request ID to correlate subsequent QuotesResponse.

Errors

Returns an error if the provided time range is invalid.

pub fn request_trades<A: DataActor>( &self, instrument_id: InstrumentId, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, limit: Option<NonZeroUsize>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>

Request historical [TradeTick] data for the given instrument_id.

Returns a unique request ID to correlate subsequent TradesResponse.

Errors

Returns an error if the provided time range is invalid.

pub fn request_bars<A: DataActor>( &self, bar_type: BarType, start: Option<DateTime<Utc>>, end: Option<DateTime<Utc>>, limit: Option<NonZeroUsize>, client_id: Option<ClientId>, params: Option<IndexMap<String, String>>, ) -> Result<UUID4>

Request historical [Bar] data for the given bar_type.

Returns a unique request ID to correlate subsequent BarsResponse.

Errors

Returns an error if the provided time range is invalid.