nautilus_common::actor::data_actor

Trait DataActor

pub trait DataActor: Actor {

    // Required method
    fn state(&self) -> ComponentState;

    // Provided methods
    fn is_ready(&self) -> bool { ... }
    fn is_running(&self) -> bool { ... }
    fn is_stopped(&self) -> bool { ... }
    fn is_disposed(&self) -> bool { ... }
    fn is_degraded(&self) -> bool { ... }
    fn is_faulted(&self) -> bool { ... }
    fn on_save(&self) -> Result<IndexMap<String, Vec<u8>>> { ... }
    fn on_load(&mut self, state: IndexMap<String, Vec<u8>>) -> Result<()> { ... }
    fn on_start(&mut self) -> Result<()> { ... }
    fn on_stop(&mut self) -> Result<()> { ... }
    fn on_resume(&mut self) -> Result<()> { ... }
    fn on_reset(&mut self) -> Result<()> { ... }
    fn on_dispose(&mut self) -> Result<()> { ... }
    fn on_degrade(&mut self) -> Result<()> { ... }
    fn on_fault(&mut self) -> Result<()> { ... }
    fn on_event(&mut self, event: &dyn Any) -> Result<()> { ... }
    fn on_time_event(&mut self, event: &TimeEvent) -> Result<()> { ... }
    fn on_data(&mut self, data: &dyn Any) -> Result<()> { ... }
    fn on_signal(&mut self, signal: &Signal) -> Result<()> { ... }
    fn on_instrument(&mut self, instrument: &InstrumentAny) -> Result<()> { ... }
    fn on_book_deltas(&mut self, deltas: &OrderBookDeltas) -> Result<()> { ... }
    fn on_book(&mut self, order_book: &OrderBook) -> Result<()> { ... }
    fn on_quote(&mut self, quote: &QuoteTick) -> Result<()> { ... }
    fn on_trade(&mut self, tick: &TradeTick) -> Result<()> { ... }
    fn on_bar(&mut self, bar: &Bar) -> Result<()> { ... }
    fn on_mark_price(&mut self, mark_price: &MarkPriceUpdate) -> Result<()> { ... }
    fn on_index_price(&mut self, index_price: &IndexPriceUpdate) -> Result<()> { ... }
    fn on_instrument_status(&mut self, data: &InstrumentStatus) -> Result<()> { ... }
    fn on_instrument_close(&mut self, update: &InstrumentClose) -> Result<()> { ... }
    fn on_historical_data(&mut self, data: &dyn Any) -> Result<()> { ... }
    fn on_historical_quotes(&mut self, quotes: &[QuoteTick]) -> Result<()> { ... }
    fn on_historical_trades(&mut self, trades: &[TradeTick]) -> Result<()> { ... }
    fn on_historical_bars(&mut self, bars: &[Bar]) -> Result<()> { ... }
    fn on_historical_mark_prices(
        &mut self,
        mark_prices: &[MarkPriceUpdate],
    ) -> Result<()> { ... }
    fn on_historical_index_prices(
        &mut self,
        index_prices: &[IndexPriceUpdate],
    ) -> Result<()> { ... }
    fn handle_data(&mut self, data: &dyn Any) { ... }
    fn handle_signal(&mut self, signal: &Signal) { ... }
    fn handle_instrument(&mut self, instrument: &InstrumentAny) { ... }
    fn handle_book_deltas(&mut self, deltas: &OrderBookDeltas) { ... }
    fn handle_book(&mut self, book: &OrderBook) { ... }
    fn handle_quote(&mut self, quote: &QuoteTick) { ... }
    fn handle_trade(&mut self, trade: &TradeTick) { ... }
    fn handle_bar(&mut self, bar: &Bar) { ... }
    fn handle_mark_price(&mut self, mark_price: &MarkPriceUpdate) { ... }
    fn handle_index_price(&mut self, index_price: &IndexPriceUpdate) { ... }
    fn handle_instrument_status(&mut self, status: &InstrumentStatus) { ... }
    fn handle_instrument_close(&mut self, close: &InstrumentClose) { ... }
    fn handle_historical_data(&mut self, data: &dyn Any) { ... }
    fn handle_time_event(&mut self, event: &TimeEvent) { ... }
    fn handle_event(&mut self, event: &dyn Any) { ... }
    fn handle_data_response(&mut self, response: &CustomDataResponse) { ... }
    fn handle_instrument_response(&mut self, response: &InstrumentResponse) { ... }
    fn handle_instruments_response(&mut self, response: &InstrumentsResponse) { ... }
    fn handle_book_response(&mut self, response: &BookResponse) { ... }
    fn handle_quotes_response(&mut self, response: &QuotesResponse) { ... }
    fn handle_trades_response(&mut self, response: &TradesResponse) { ... }
    fn handle_bars_response(&mut self, response: &BarsResponse) { ... }

}

Required Methods

fn state(&self) -> ComponentState

Returns the ComponentState of the actor.

Provided Methods

fn is_ready(&self) -> bool

Returns true if the actor is in a Ready state.

fn is_running(&self) -> bool

Returns true if the actor is in a Running state.

fn is_stopped(&self) -> bool

Returns true if the actor is in a Stopped state.

fn is_disposed(&self) -> bool

Returns true if the actor is in a Disposed state.

fn is_degraded(&self) -> bool

Returns true if the actor is in a Degraded state.

fn is_faulted(&self) -> bool

Returns true if the actor is in a Faulted state.

fn on_save(&self) -> Result<IndexMap<String, Vec<u8>>>

Actions to be performed when the actor state is saved.

Errors

Returns an error if saving the actor state fails.

fn on_load(&mut self, state: IndexMap<String, Vec<u8>>) -> Result<()>

Actions to be performed when the actor state is loaded.

Errors

Returns an error if loading the actor state fails.

fn on_start(&mut self) -> Result<()>

Actions to be performed on start.

Errors

Returns an error if starting the actor fails.

fn on_stop(&mut self) -> Result<()>

Actions to be performed on stop.

Errors

Returns an error if stopping the actor fails.

fn on_resume(&mut self) -> Result<()>

Actions to be performed on resume.

Errors

Returns an error if resuming the actor fails.

fn on_reset(&mut self) -> Result<()>

Actions to be performed on reset.

Errors

Returns an error if resetting the actor fails.

fn on_dispose(&mut self) -> Result<()>

Actions to be performed on dispose.

Errors

Returns an error if disposing the actor fails.

fn on_degrade(&mut self) -> Result<()>

Actions to be performed on degrade.

Errors

Returns an error if degrading the actor fails.

fn on_fault(&mut self) -> Result<()>

Actions to be performed on fault.

Errors

Returns an error if faulting the actor fails.

fn on_event(&mut self, event: &dyn Any) -> Result<()>

Actions to be performed when receiving an event.

Errors

Returns an error if handling the event fails.

fn on_time_event(&mut self, event: &TimeEvent) -> Result<()>

Actions to be performed when receiving a time event.

Errors

Returns an error if handling the time event fails.

fn on_data(&mut self, data: &dyn Any) -> Result<()>

Actions to be performed when receiving custom data.

Errors

Returns an error if handling the data fails.

fn on_signal(&mut self, signal: &Signal) -> Result<()>

Actions to be performed when receiving a signal.

Errors

Returns an error if handling the signal fails.

fn on_instrument(&mut self, instrument: &InstrumentAny) -> Result<()>

Actions to be performed when receiving an instrument.

Errors

Returns an error if handling the instrument fails.

fn on_book_deltas(&mut self, deltas: &OrderBookDeltas) -> Result<()>

Actions to be performed when receiving order book deltas.

Errors

Returns an error if handling the book deltas fails.

fn on_book(&mut self, order_book: &OrderBook) -> Result<()>

Actions to be performed when receiving an order book.

Errors

Returns an error if handling the book fails.

fn on_quote(&mut self, quote: &QuoteTick) -> Result<()>

Actions to be performed when receiving a quote.

Errors

Returns an error if handling the quote fails.

fn on_trade(&mut self, tick: &TradeTick) -> Result<()>

Actions to be performed when receiving a trade.

Errors

Returns an error if handling the trade fails.

fn on_bar(&mut self, bar: &Bar) -> Result<()>

Actions to be performed when receiving a bar.

Errors

Returns an error if handling the bar fails.

fn on_mark_price(&mut self, mark_price: &MarkPriceUpdate) -> Result<()>

Actions to be performed when receiving a mark price update.

Errors

Returns an error if handling the mark price update fails.

fn on_index_price(&mut self, index_price: &IndexPriceUpdate) -> Result<()>

Actions to be performed when receiving an index price update.

Errors

Returns an error if handling the index price update fails.

fn on_instrument_status(&mut self, data: &InstrumentStatus) -> Result<()>

Actions to be performed when receiving an instrument status update.

Errors

Returns an error if handling the instrument status update fails.

fn on_instrument_close(&mut self, update: &InstrumentClose) -> Result<()>

Actions to be performed when receiving an instrument close update.

Errors

Returns an error if handling the instrument close update fails.

fn on_historical_data(&mut self, data: &dyn Any) -> Result<()>

Actions to be performed when receiving historical data.

Errors

Returns an error if handling the historical data fails.

fn on_historical_quotes(&mut self, quotes: &[QuoteTick]) -> Result<()>

Actions to be performed when receiving historical quotes.

Errors

Returns an error if handling the historical quotes fails.

fn on_historical_trades(&mut self, trades: &[TradeTick]) -> Result<()>

Actions to be performed when receiving historical trades.

Errors

Returns an error if handling the historical trades fails.

fn on_historical_bars(&mut self, bars: &[Bar]) -> Result<()>

Actions to be performed when receiving historical bars.

Errors

Returns an error if handling the historical bars fails.

fn on_historical_mark_prices( &mut self, mark_prices: &[MarkPriceUpdate], ) -> Result<()>

Actions to be performed when receiving historical mark prices.

Errors

Returns an error if handling the historical mark prices fails.

fn on_historical_index_prices( &mut self, index_prices: &[IndexPriceUpdate], ) -> Result<()>

Actions to be performed when receiving historical index prices.

Errors

Returns an error if handling the historical index prices fails.

fn handle_data(&mut self, data: &dyn Any)

Handles a received custom data point.

fn handle_signal(&mut self, signal: &Signal)

Handles a received signal.

fn handle_instrument(&mut self, instrument: &InstrumentAny)

Handles a received instrument.

fn handle_book_deltas(&mut self, deltas: &OrderBookDeltas)

Handles received order book deltas.

fn handle_book(&mut self, book: &OrderBook)

Handles a received order book reference.

fn handle_quote(&mut self, quote: &QuoteTick)

Handles a received quote.

fn handle_trade(&mut self, trade: &TradeTick)

Handles a received trade.

fn handle_bar(&mut self, bar: &Bar)

Handles a receiving bar.

fn handle_mark_price(&mut self, mark_price: &MarkPriceUpdate)

Handles a received mark price update.

fn handle_index_price(&mut self, index_price: &IndexPriceUpdate)

Handles a received index price update.

fn handle_instrument_status(&mut self, status: &InstrumentStatus)

Handles a received instrument status.

fn handle_instrument_close(&mut self, close: &InstrumentClose)

Handles a received instrument close.

fn handle_historical_data(&mut self, data: &dyn Any)

Handles received historical data.

fn handle_time_event(&mut self, event: &TimeEvent)

Handles a received time event.

fn handle_event(&mut self, event: &dyn Any)

Handles a received event.

fn handle_data_response(&mut self, response: &CustomDataResponse)

Handles a data response.

fn handle_instrument_response(&mut self, response: &InstrumentResponse)

Handles an instrument response.

fn handle_instruments_response(&mut self, response: &InstrumentsResponse)

Handles an instruments response.

fn handle_book_response(&mut self, response: &BookResponse)

Handles a book response.

fn handle_quotes_response(&mut self, response: &QuotesResponse)

Handles a quotes response.

fn handle_trades_response(&mut self, response: &TradesResponse)

Handles a trades response.

fn handle_bars_response(&mut self, response: &BarsResponse)

Handles a bars response.