Struct nautilus_common::cache::core::Cache

pub struct Cache { /* private fields */ }

A common in-memory Cache for market and execution related data.

Implementations

impl Cache

pub fn new(config: CacheConfig, database: Option<CacheDatabaseAdapter>) -> Self

pub fn cache_general(&mut self) -> Result<()>

Clear the current general cache and load the general objects from the cache database.

pub fn cache_currencies(&mut self) -> Result<()>

Clear the current currencies cache and load currencies from the cache database.

pub fn cache_instruments(&mut self) -> Result<()>

Clear the current instruments cache and load instruments from the cache database.

pub fn cache_synthetics(&mut self) -> Result<()>

Clear the current synthetic instruments cache and load synthetic instruments from the cache database.

pub fn cache_accounts(&mut self) -> Result<()>

Clear the current accounts cache and load accounts from the cache database.

pub fn cache_orders(&mut self) -> Result<()>

Clear the current orders cache and load orders from the cache database.

pub fn cache_positions(&mut self) -> Result<()>

Clear the current positions cache and load positions from the cache database.

pub fn build_index(&mut self)

Clear the current cache index and re-build.

pub fn check_residuals(&self) -> bool

Check for any residual open state and log warnings if any are found.

‘Open state’ is considered to be open orders and open positions.

pub fn clear_index(&mut self)

Clear the caches index.

pub fn reset(&mut self)

Reset the cache.

All stateful fields are reset to their initial value.

pub fn dispose(&self) -> Result<()>

Dispose of the cache which will close any underlying database adapter.

pub fn flush_db(&self) -> Result<()>

Flush the caches database which permanently removes all persisted data.

pub fn add(&mut self, key: &str, value: Vec<u8>) -> Result<()>

Add the given general object to the cache.

The cache is agnostic to what the object actually is (and how it may be serialized), offering maximum flexibility.

pub fn add_order_book(&mut self, book: OrderBook) -> Result<()>

Add the given order book to the cache.

pub fn add_quote(&mut self, quote: QuoteTick) -> Result<()>

Add the given quote tick to the cache.

pub fn add_quotes(&mut self, quotes: &[QuoteTick]) -> Result<()>

Add the given quotes to the cache.

pub fn add_trade(&mut self, trade: TradeTick) -> Result<()>

Add the given trade tick to the cache.

pub fn add_trades(&mut self, trades: &[TradeTick]) -> Result<()>

Add the give trades to the cache.

pub fn add_bar(&mut self, bar: Bar) -> Result<()>

Add the given bar to the cache.

pub fn add_bars(&mut self, bars: &[Bar]) -> Result<()>

Add the given bars to the cache.

pub fn add_currency(&mut self, currency: Currency) -> Result<()>

Add the given currency to the cache.

pub fn add_instrument(&mut self, instrument: InstrumentAny) -> Result<()>

Add the given instrument to the cache.

pub fn add_synthetic(&mut self, synthetic: SyntheticInstrument) -> Result<()>

Add the given synthetic instrument to the cache.

pub fn add_account(&mut self, account: Box<dyn Account>) -> Result<()>

Add the given account to the cache.

pub fn add_order( &mut self, order: OrderAny, position_id: Option<PositionId>, client_id: Option<ClientId>, replace_existing: bool ) -> Result<()>

Add the order to the cache indexed with any given identifiers.

Parameters

override_existing: If the added order should ‘override’ any existing order and replace it in the cache. This is currently used for emulated orders which are being released and transformed into another type.

Errors

If not replace_existing and the order.client_order_id is already contained in the cache.

pub fn add_position_id( &mut self, position_id: &PositionId, venue: &Venue, client_order_id: &ClientOrderId, strategy_id: &StrategyId ) -> Result<()>

Index the given position_id with the other given IDs.

pub fn add_position( &mut self, position: Position, oms_type: OmsType ) -> Result<()>

pub fn update_account(&mut self, account: &dyn Account) -> Result<()>

Update the given account in the cache.

pub fn update_order(&mut self, order: &OrderAny) -> Result<()>

Update the given order in the cache.

pub fn update_order_pending_cancel_local(&mut self, order: &OrderAny)

Update the given order as pending cancel locally.

pub fn update_position(&mut self, position: &Position) -> Result<()>

Update the given position in the cache.

pub fn client_order_ids( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId> ) -> HashSet<ClientOrderId>

pub fn client_order_ids_open( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId> ) -> HashSet<ClientOrderId>

pub fn client_order_ids_closed( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId> ) -> HashSet<ClientOrderId>

pub fn client_order_ids_emulated( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId> ) -> HashSet<ClientOrderId>

pub fn client_order_ids_inflight( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId> ) -> HashSet<ClientOrderId>

pub fn position_ids( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId> ) -> HashSet<PositionId>

pub fn position_open_ids( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId> ) -> HashSet<PositionId>

pub fn position_closed_ids( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId> ) -> HashSet<PositionId>

pub fn actor_ids(&self) -> HashSet<ComponentId>

pub fn strategy_ids(&self) -> HashSet<StrategyId>

pub fn exec_algorithm_ids(&self) -> HashSet<ExecAlgorithmId>

pub fn order(&self, client_order_id: &ClientOrderId) -> Option<&OrderAny>

pub fn client_order_id( &self, venue_order_id: &VenueOrderId ) -> Option<&ClientOrderId>

pub fn venue_order_id( &self, client_order_id: &ClientOrderId ) -> Option<VenueOrderId>

pub fn client_id(&self, client_order_id: &ClientOrderId) -> Option<&ClientId>

pub fn orders( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<OrderSide> ) -> Vec<&OrderAny>

pub fn orders_open( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<OrderSide> ) -> Vec<&OrderAny>

pub fn orders_closed( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<OrderSide> ) -> Vec<&OrderAny>

pub fn orders_emulated( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<OrderSide> ) -> Vec<&OrderAny>

pub fn orders_inflight( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<OrderSide> ) -> Vec<&OrderAny>

pub fn orders_for_position(&self, position_id: &PositionId) -> Vec<&OrderAny>

pub fn order_exists(&self, client_order_id: &ClientOrderId) -> bool

pub fn is_order_open(&self, client_order_id: &ClientOrderId) -> bool

pub fn is_order_closed(&self, client_order_id: &ClientOrderId) -> bool

pub fn is_order_emulated(&self, client_order_id: &ClientOrderId) -> bool

pub fn is_order_inflight(&self, client_order_id: &ClientOrderId) -> bool

pub fn is_order_pending_cancel_local( &self, client_order_id: &ClientOrderId ) -> bool

pub fn orders_open_count( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<OrderSide> ) -> usize

pub fn orders_closed_count( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<OrderSide> ) -> usize

pub fn orders_emulated_count( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<OrderSide> ) -> usize

pub fn orders_inflight_count( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<OrderSide> ) -> usize

pub fn orders_total_count( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<OrderSide> ) -> usize

pub fn order_list(&self, order_list_id: &OrderListId) -> Option<&OrderList>

pub fn order_lists( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId> ) -> Vec<&OrderList>

pub fn order_list_exists(&self, order_list_id: &OrderListId) -> bool

pub fn orders_for_exec_algorithm( &self, exec_algorithm_id: &ExecAlgorithmId, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<OrderSide> ) -> Vec<&OrderAny>

pub fn orders_for_exec_spawn( &self, exec_spawn_id: &ClientOrderId ) -> Vec<&OrderAny>

pub fn exec_spawn_total_quantity( &self, exec_spawn_id: &ClientOrderId, active_only: bool ) -> Option<Quantity>

pub fn exec_spawn_total_filled_qty( &self, exec_spawn_id: &ClientOrderId, active_only: bool ) -> Option<Quantity>

pub fn exec_spawn_total_leaves_qty( &self, exec_spawn_id: &ClientOrderId, active_only: bool ) -> Option<Quantity>

pub fn position(&self, position_id: &PositionId) -> Option<&Position>

pub fn position_for_order( &self, client_order_id: &ClientOrderId ) -> Option<&Position>

pub fn position_id( &self, client_order_id: &ClientOrderId ) -> Option<&PositionId>

pub fn positions( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<PositionSide> ) -> Vec<&Position>

pub fn positions_open( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<PositionSide> ) -> Vec<&Position>

pub fn positions_closed( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<PositionSide> ) -> Vec<&Position>

pub fn position_exists(&self, position_id: &PositionId) -> bool

pub fn is_position_open(&self, position_id: &PositionId) -> bool

pub fn is_position_closed(&self, position_id: &PositionId) -> bool

pub fn positions_open_count( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<PositionSide> ) -> u64

pub fn positions_closed_count( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<PositionSide> ) -> u64

pub fn positions_total_count( &self, venue: Option<&Venue>, instrument_id: Option<&InstrumentId>, strategy_id: Option<&StrategyId>, side: Option<PositionSide> ) -> u64

pub fn strategy_id_for_order( &self, client_order_id: &ClientOrderId ) -> Option<&StrategyId>

pub fn strategy_id_for_position( &self, position_id: &PositionId ) -> Option<&StrategyId>

pub fn get(&self, key: &str) -> Result<Option<&[u8]>>

pub fn price( &self, instrument_id: &InstrumentId, price_type: PriceType ) -> Option<Price>

pub fn quote_ticks( &self, instrument_id: &InstrumentId ) -> Option<Vec<QuoteTick>>

pub fn trade_ticks( &self, instrument_id: &InstrumentId ) -> Option<Vec<TradeTick>>

pub fn bars(&self, bar_type: &BarType) -> Option<Vec<Bar>>

pub fn order_book(&self, instrument_id: &InstrumentId) -> Option<&OrderBook>

pub fn quote_tick(&self, instrument_id: &InstrumentId) -> Option<&QuoteTick>

pub fn trade_tick(&self, instrument_id: &InstrumentId) -> Option<&TradeTick>

pub fn bar(&self, bar_type: &BarType) -> Option<&Bar>

pub fn book_update_count(&self, instrument_id: &InstrumentId) -> u64

pub fn quote_tick_count(&self, instrument_id: &InstrumentId) -> u64

pub fn trade_tick_count(&self, instrument_id: &InstrumentId) -> u64

pub fn bar_count(&self, bar_type: &BarType) -> u64

pub fn has_order_book(&self, instrument_id: &InstrumentId) -> bool

pub fn has_quote_ticks(&self, instrument_id: &InstrumentId) -> bool

pub fn has_trade_ticks(&self, instrument_id: &InstrumentId) -> bool

pub fn has_bars(&self, bar_type: &BarType) -> bool

pub fn instrument(&self, instrument_id: &InstrumentId) -> Option<&InstrumentAny>

pub fn instrument_ids(&self, venue: &Venue) -> Vec<&InstrumentId>

pub fn instruments(&self, venue: &Venue) -> Vec<&InstrumentAny>

pub fn bar_types( &self, instrument_id: Option<&InstrumentId>, price_type: Option<&PriceType>, aggregation_source: AggregationSource ) -> Vec<&BarType>

pub fn synthetic( &self, instrument_id: &InstrumentId ) -> Option<&SyntheticInstrument>

pub fn synthetic_ids(&self) -> Vec<&InstrumentId>

pub fn synthetics(&self) -> Vec<&SyntheticInstrument>

pub fn account(&self, account_id: &AccountId) -> Option<&dyn Account>

pub fn account_for_venue(&self, venue: &Venue) -> Option<&dyn Account>

pub fn account_id(&self, venue: &Venue) -> Option<&AccountId>

pub fn accounts(&self, account_id: &AccountId) -> Vec<&dyn Account>