Struct nautilus_common::cache::Cache

pub struct Cache { /* private fields */ }

Implementations

impl Cache

pub fn new(config: CacheConfig, database: Option<CacheDatabaseAdapter>) -> Self

pub fn cache_general(&mut self) -> Result<()>

pub fn cache_currencies(&mut self) -> Result<()>

pub fn cache_instruments(&mut self) -> Result<()>

pub fn cache_synthetics(&mut self) -> Result<()>

pub fn cache_accounts(&mut self) -> Result<()>

pub fn cache_orders(&mut self) -> Result<()>

pub fn cache_positions(&mut self) -> Result<()>

pub fn build_index(&self)

pub fn check_integrity(&self) -> bool

pub fn check_residuals(&self)

pub fn clear_index(&mut self)

pub fn reset(&mut self)

Reset the cache.

All stateful fields are reset to their initial value.

pub fn dispose(&self) -> Result<()>

pub fn flush_db(&self) -> Result<()>

pub fn add(&mut self, key: &str, value: Vec<u8>) -> Result<()>

pub fn add_order_book(&mut self, book: OrderBook) -> Result<()>

pub fn add_quote(&mut self, quote: QuoteTick) -> Result<()>

pub fn add_quotes(&mut self, quotes: &[QuoteTick]) -> Result<()>

pub fn add_trade(&mut self, trade: TradeTick) -> Result<()>

pub fn add_trades(&mut self, trades: &[TradeTick]) -> Result<()>

pub fn add_bar(&mut self, bar: Bar) -> Result<()>

pub fn add_bars(&mut self, bars: &[Bar]) -> Result<()>

pub fn add_currency(&mut self, currency: Currency) -> Result<()>

pub fn add_instrument(&mut self, instrument: InstrumentAny) -> Result<()>

pub fn add_synthetic(&mut self, synthetic: SyntheticInstrument) -> Result<()>

pub fn add_account(&mut self, account: Box<dyn Account>) -> Result<()>

pub fn add_order( &mut self, order: OrderAny, _position_id: Option<PositionId>, client_id: Option<ClientId>, replace_existing: bool ) -> Result<()>

Add the order to the cache indexed with any given identifiers.

Parameters

override_existing: If the added order should ‘override’ any existing order and replace it in the cache. This is currently used for emulated orders which are being released and transformed into another type.

Errors

If not replace_existing and the order.client_order_id is already contained in the cache.

pub fn client_order_ids( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId> ) -> HashSet<ClientOrderId>

pub fn client_order_ids_open( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId> ) -> HashSet<ClientOrderId>

pub fn client_order_ids_closed( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId> ) -> HashSet<ClientOrderId>

pub fn client_order_ids_emulated( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId> ) -> HashSet<ClientOrderId>

pub fn client_order_ids_inflight( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId> ) -> HashSet<ClientOrderId>

pub fn position_ids( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId> ) -> HashSet<PositionId>

pub fn position_open_ids( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId> ) -> HashSet<PositionId>

pub fn position_closed_ids( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId> ) -> HashSet<PositionId>

pub fn actor_ids(&self) -> HashSet<ComponentId>

pub fn strategy_ids(&self) -> HashSet<StrategyId>

pub fn exec_algorithm_ids(&self) -> HashSet<ExecAlgorithmId>

pub fn order(&self, client_order_id: ClientOrderId) -> Option<&OrderAny>

pub fn client_order_id( &self, venue_order_id: VenueOrderId ) -> Option<&ClientOrderId>

pub fn venue_order_id( &self, client_order_id: ClientOrderId ) -> Option<VenueOrderId>

pub fn client_id(&self, client_order_id: ClientOrderId) -> Option<&ClientId>

pub fn orders( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId>, side: Option<OrderSide> ) -> Vec<&OrderAny>

pub fn orders_open( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId>, side: Option<OrderSide> ) -> Vec<&OrderAny>

pub fn orders_closed( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId>, side: Option<OrderSide> ) -> Vec<&OrderAny>

pub fn orders_emulated( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId>, side: Option<OrderSide> ) -> Vec<&OrderAny>

pub fn orders_inflight( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId>, side: Option<OrderSide> ) -> Vec<&OrderAny>

pub fn orders_for_position(&self, position_id: PositionId) -> Vec<&OrderAny>

pub fn order_exists(&self, client_order_id: ClientOrderId) -> bool

pub fn is_order_open(&self, client_order_id: ClientOrderId) -> bool

pub fn is_order_closed(&self, client_order_id: ClientOrderId) -> bool

pub fn is_order_emulated(&self, client_order_id: ClientOrderId) -> bool

pub fn is_order_inflight(&self, client_order_id: ClientOrderId) -> bool

pub fn is_order_pending_cancel_local( &self, client_order_id: ClientOrderId ) -> bool

pub fn orders_open_count( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId>, side: Option<OrderSide> ) -> usize

pub fn orders_closed_count( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId>, side: Option<OrderSide> ) -> usize

pub fn orders_emulated_count( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId>, side: Option<OrderSide> ) -> usize

pub fn orders_inflight_count( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId>, side: Option<OrderSide> ) -> usize

pub fn orders_total_count( &self, venue: Option<Venue>, instrument_id: Option<InstrumentId>, strategy_id: Option<StrategyId>, side: Option<OrderSide> ) -> usize

pub fn get(&self, key: &str) -> Result<Option<&Vec<u8>>>