Trait nautilus_common::clock::Clock

pub trait Clock {
    // Required methods
    fn timer_names(&self) -> Vec<&str>;
    fn timer_count(&self) -> usize;
    fn register_default_handler(&mut self, callback: EventHandler);
    fn set_time_alert_ns(
        &mut self,
        name: &str,
        alert_time_ns: UnixNanos,
        callback: Option<EventHandler>
    );
    fn set_timer_ns(
        &mut self,
        name: &str,
        interval_ns: u64,
        start_time_ns: UnixNanos,
        stop_time_ns: Option<UnixNanos>,
        callback: Option<EventHandler>
    );
    fn next_time_ns(&self, name: &str) -> UnixNanos;
    fn cancel_timer(&mut self, name: &str);
    fn cancel_timers(&mut self);
}

Represents a type of clock.

Notes

An active timer is one which has not expired (timer.is_expired == False).

Required Methods

fn timer_names(&self) -> Vec<&str>

Return the names of active timers in the clock.

fn timer_count(&self) -> usize

Return the count of active timers in the clock.

fn register_default_handler(&mut self, callback: EventHandler)

Register a default event handler for the clock. If a Timer does not have an event handler, then this handler is used.

fn set_time_alert_ns( &mut self, name: &str, alert_time_ns: UnixNanos, callback: Option<EventHandler> )

Set a Timer to alert at a particular time. Optional callback gets used to handle generated events.

fn set_timer_ns( &mut self, name: &str, interval_ns: u64, start_time_ns: UnixNanos, stop_time_ns: Option<UnixNanos>, callback: Option<EventHandler> )

Set a Timer to start alerting at every interval between start and stop time. Optional callback gets used to handle generated event.

fn next_time_ns(&self, name: &str) -> UnixNanos

fn cancel_timer(&mut self, name: &str)

fn cancel_timers(&mut self)