nautilus_common::component

Trait Component

pub trait Component {

    // Required methods
    fn component_id(&self) -> ComponentId;
    fn state(&self) -> ComponentState;
    fn transition_state(&mut self, trigger: ComponentTrigger) -> Result<()>;
    fn register(
        &mut self,
        trader_id: TraderId,
        clock: Rc<RefCell<dyn Clock>>,
        cache: Rc<RefCell<Cache>>,
    ) -> Result<()>;

    // Provided methods
    fn is_ready(&self) -> bool { ... }
    fn not_running(&self) -> bool { ... }
    fn is_running(&self) -> bool { ... }
    fn is_stopped(&self) -> bool { ... }
    fn is_degraded(&self) -> bool { ... }
    fn is_faulted(&self) -> bool { ... }
    fn is_disposed(&self) -> bool { ... }
    fn initialize(&mut self) -> Result<()> { ... }
    fn start(&mut self) -> Result<()> { ... }
    fn stop(&mut self) -> Result<()> { ... }
    fn resume(&mut self) -> Result<()> { ... }
    fn degrade(&mut self) -> Result<()> { ... }
    fn fault(&mut self) -> Result<()> { ... }
    fn reset(&mut self) -> Result<()> { ... }
    fn dispose(&mut self) -> Result<()> { ... }
    fn on_start(&mut self) -> Result<()> { ... }
    fn on_stop(&mut self) -> Result<()> { ... }
    fn on_resume(&mut self) -> Result<()> { ... }
    fn on_reset(&mut self) -> Result<()> { ... }
    fn on_dispose(&mut self) -> Result<()> { ... }
    fn on_degrade(&mut self) -> Result<()> { ... }
    fn on_fault(&mut self) -> Result<()> { ... }

}

Components have state and lifecycle management capabilities.

Required Methods

fn component_id(&self) -> ComponentId

Returns the unique identifier for this component.

fn state(&self) -> ComponentState

Returns the current state of the component.

fn transition_state(&mut self, trigger: ComponentTrigger) -> Result<()>

Transition the component with the state trigger.

Errors

Returns an error if the trigger is an invalid transition from the current state.

fn register( &mut self, trader_id: TraderId, clock: Rc<RefCell<dyn Clock>>, cache: Rc<RefCell<Cache>>, ) -> Result<()>

Registers the component with a system.

Errors

Returns an error if the component fails to register.

Provided Methods

fn is_ready(&self) -> bool

Returns whether the component is ready.

fn not_running(&self) -> bool

Returns whether the component is not running.

fn is_running(&self) -> bool

Returns whether the component is running.

fn is_stopped(&self) -> bool

Returns whether the component is stopped.

fn is_degraded(&self) -> bool

Returns whether the component has been degraded.

fn is_faulted(&self) -> bool

Returns whether the component has been faulted.

fn is_disposed(&self) -> bool

Returns whether the component has been disposed.

fn initialize(&mut self) -> Result<()>

Initializes the component.

Errors

Returns an error if the initialization state transition fails.

fn start(&mut self) -> Result<()>

Starts the component.

Errors

Returns an error if the component fails to start.

fn stop(&mut self) -> Result<()>

Stops the component.

Errors

Returns an error if the component fails to stop.

fn resume(&mut self) -> Result<()>

Resumes the component.

Errors

Returns an error if the component fails to resume.

fn degrade(&mut self) -> Result<()>

Degrades the component.

Errors

Returns an error if the component fails to degrade.

fn fault(&mut self) -> Result<()>

Faults the component.

Errors

Returns an error if the component fails to fault.

fn reset(&mut self) -> Result<()>

Resets the component to its initial state.

Errors

Returns an error if the component fails to reset.

fn dispose(&mut self) -> Result<()>

Disposes of the component, releasing any resources.

Errors

Returns an error if the component fails to dispose.

fn on_start(&mut self) -> Result<()>

Actions to be performed on start.

Errors

Returns an error if starting the actor fails.

fn on_stop(&mut self) -> Result<()>

Actions to be performed on stop.

Errors

Returns an error if stopping the actor fails.

fn on_resume(&mut self) -> Result<()>

Actions to be performed on resume.

Errors

Returns an error if resuming the actor fails.

fn on_reset(&mut self) -> Result<()>

Actions to be performed on reset.

Errors

Returns an error if resetting the actor fails.

fn on_dispose(&mut self) -> Result<()>

Actions to be performed on dispose.

Errors

Returns an error if disposing the actor fails.

fn on_degrade(&mut self) -> Result<()>

Actions to be performed on degrade.

Errors

Returns an error if degrading the actor fails.

fn on_fault(&mut self) -> Result<()>

Actions to be performed on fault.

Errors

Returns an error if faulting the actor fails.