nautilus_common::component

Trait Component

pub trait Component: Debug {
    // Required methods
    fn id(&self) -> ComponentId;
    fn state(&self) -> ComponentState;
    fn trigger(&self) -> ComponentTrigger;
    fn is_running(&self) -> bool;
    fn is_stopped(&self) -> bool;
    fn is_disposed(&self) -> bool;
    fn start(&mut self) -> Result<()>;
    fn stop(&mut self) -> Result<()>;
    fn reset(&mut self) -> Result<()>;
    fn dispose(&mut self) -> Result<()>;
    fn handle_event(&mut self, event: TimeEvent);
}

Common trait for components.

Required Methods

fn id(&self) -> ComponentId

Returns the unique identifier for this component.

fn state(&self) -> ComponentState

Returns the current state of the component.

fn trigger(&self) -> ComponentTrigger

Returns the component trigger.

fn is_running(&self) -> bool

Returns whether the component is currently running.

fn is_stopped(&self) -> bool

Returns whether the component is stopped.

fn is_disposed(&self) -> bool

Returns whether the component has been disposed.

fn start(&mut self) -> Result<()>

Starts the component.

Errors

Returns an error if the component fails to start.

fn stop(&mut self) -> Result<()>

Stops the component.

Errors

Returns an error if the component fails to stop.

fn reset(&mut self) -> Result<()>

Resets the component to its initial state.

Errors

Returns an error if the component fails to reset.

fn dispose(&mut self) -> Result<()>

Disposes of the component, releasing any resources.

Errors

Returns an error if the component fails to dispose.

fn handle_event(&mut self, event: TimeEvent)

Handles a timer event (TBD).