Trait nautilus_common::msgbus::database::MessageBusDatabaseAdapter

pub trait MessageBusDatabaseAdapter {
    type DatabaseType;

    // Required methods
    fn new(
        trader_id: TraderId,
        instance_id: UUID4,
        config: HashMap<String, Value>
    ) -> Result<Self::DatabaseType>;
    fn publish(&self, topic: String, payload: Vec<u8>) -> Result<()>;
    fn close(&mut self) -> Result<()>;
}

Provides a generic message bus database facade.

The main operations take a consistent key and payload which should provide enough information to implement the message bus database in many different technologies.

Delete operations may need a payload to target specific values.

Required Associated Types

type DatabaseType

Required Methods

fn new( trader_id: TraderId, instance_id: UUID4, config: HashMap<String, Value> ) -> Result<Self::DatabaseType>

fn publish(&self, topic: String, payload: Vec<u8>) -> Result<()>

fn close(&mut self) -> Result<()>

Object Safety

This trait is not object safe.